bitfield,  file checking and directory creation
bitfields for all peers initialized to 0 or 1 if they have the entire
file or not.

If peer has entire file, it checks to make sure file is in the correct
directory e.g. ../peer_id/TheFile.dat

Peers that dont have entire file create directory to store the
downloaded files ../peer_id/